requirements.txt
nomeroff-net>=4.0.0
opencv-python>=4.7.0
numpy>=1.21.0
PyYAML>=6.0
pytest>=7.3.1
tqdm>=4.65.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0